Fix EF.SQlServer reference and JSON versions
The EntityFramework.SqlServer reference in the Core project was pointing
to the wrong directory. I fixed the hint path to point to the correct
location.

The Newtonsoft.JSON package in the Test project was several versions
behind the Web project. Performed an Update-Package to match verion with
other projects.

Code now builds with no error or warnings.